+++ changelog.md
@@ -1,3 +1,9 @@
+### v2.0.1
+
+## Changes 
+
+Updated toc for patch 9.0.5.
+
 ### v2.0.0
 ---
 * Updated for Shadowlands and RCLootCouncil v3.0.